+++ app/[language]/[slug]/page.tsx
@@ -1,7 +1,6 @@
 import { languages } from "@/config/languages";
 import { Language } from "@/types/language";
 import { getSlugs, loadPageMdx } from "@/helpers/content-helper";
-import type { Metadata } from "next";
 
 export default async function Page({
   params,
@@ -9,10 +8,7 @@ export default async function Page({
   params: Promise<{ language: Language; slug: string }>;
 }) {
   const { language, slug } = await params;
-  const { default: PageContent, frontmatter } = await loadPageMdx(
-    language,
-    slug
-  );
+  const { default: PageContent } = await loadPageMdx(language, slug);
   return <PageContent />;
 }
 
@@ -27,37 +23,6 @@ export async function generateStaticParams() {
         };
       })
   );
-}
-
-export async function generateMetadata({
-  params,
-}: {
-  params: Promise<{ language: Language; slug: string }>;
-}): Promise<Metadata> {
-  const { language, slug } = await params;
-  const { frontmatter } = await loadPageMdx(language, slug);
-  return {
-    title: frontmatter.title,
-    description: frontmatter.description,
-    openGraph: {
-      url: `/${language}/${slug}`,
-    },
-    alternates: {
-      canonical: `/${language}/${slug}`,
-      languages: Object.fromEntries(
-        languages.map((lang) => [
-          lang,
-          `/${lang}/${lang === language ? slug : frontmatter[lang] || ""}`,
-        ])
-      ),
-    },
-    robots: {
-      index: frontmatter.disableRobots === true ? false : true,
-      googleBot: {
-        index: frontmatter.disableRobots === true ? false : true,
-      },
-    },
-  };
 }
 
 export const dynamicParams = false;
