+++ content/pages/en/index.mdx
@@ -1,5 +1,5 @@
 ---
-title: Hjem
+title: Home
 description: Queer Hangout er en møteplass for voksne skeive i Drammen. Vi er en frivillig organisasjon som møtes 1 gang i måneden. Vi jobber mot ensomhet i det skeive miljøet.
 menu: 0
 ---
