+++ helpers/content-helper.ts
@@ -1,8 +1,8 @@
-import { Language } from "@/types/language";
+import { Language, Translated } from "@/types/language";
 import fs from "node:fs";
 import path from "node:path";
 import { MenuItem, PageMdx } from "@/types/content";
-import { describe } from "node:test";
+import { languages } from "@/config/languages";
 
 export const getMdxFiles = (language: Language) =>
   fs.readdirSync(path.join("content", "pages", language));
@@ -35,3 +35,22 @@ export const loadMenuItems = async (language: Language) =>
       )
     ).filter((a) => a.index != null) as MenuItem[]
   ).toSorted((a, b) => a.index - b.index);
+
+export const getAlternates = async (
+  language: Language,
+  slug: string
+): Promise<Translated<string>> =>
+  Object.fromEntries(
+    await Promise.all(
+      languages.map(async (lang) => {
+        if (lang === language) return [lang, `/${lang}/${slug}`];
+        else
+          return [
+            lang,
+            `/${lang}/${
+              (await loadPageMdx(language, slug)).frontmatter[lang] || ""
+            }`,
+          ];
+      })
+    )
+  ) as Translated<string>;
